+++ lib/server/index.js
@@ -18,7 +18,7 @@ import {
   keywords,
   copyright,
   robots
-} from 'helpers/const/setting-keys';
+} from 'statics/settings-keys';
 
 import config from '../../config';
 import middleware from './middleware';
