+++ lib/server/routers/public.js
@@ -17,7 +17,7 @@ import {
   mailUser,
   mailPass,
   mailTo
-} from 'helpers/const/setting-keys';
+} from 'statics/settings-keys';
 
 import MediaModel from '../models/media';
 import schema from '../schema';
