+++ lib/shared/screens/admin/screens/settings/screens/analytics/index.js
@@ -3,7 +3,7 @@ import Content from 'components/content';
 import React from 'react';
 import SettingsForm from 'components/settings-form';
 import mapSettingsIds from 'helpers/map-settings-ids';
-import {googleAnalytics} from 'helpers/const/setting-keys';
+import {googleAnalytics} from 'statics/settings-keys';
 const options = [
   {
     label: 'Google analytics tracking ID',
