+++ lib/shared/screens/admin/screens/settings/screens/general/index.js
@@ -3,7 +3,7 @@ import Content from 'components/content';
 import React from 'react';
 import SettingsForm from 'components/settings-form';
 import mapSettingsIds from 'helpers/map-settings-ids';
-import {title, frontpage, favicon, webclip} from 'helpers/const/setting-keys';
+import {title, frontpage, favicon, webclip} from 'statics/settings-keys';
 const options = [
   {
     label: 'Site Title',
