+++ lib/shared/screens/admin/screens/settings/screens/seo/index.js
@@ -9,7 +9,7 @@ import {
   keywords,
   robots,
   dctitle
-} from 'helpers/const/setting-keys';
+} from 'statics/settings-keys';
 
 const options = [
   {
