+++ lib/shared/statics/settings-keys.js
